+++ src/utils/NetworkingUtils.js
@@ -4,7 +4,7 @@ import store from "../redux/store/store";
 import { navigate } from "./NavigationUtils";
 
 const HopPassionClient = axios.create({
-  baseURL: "https://hoppassion-server.1.ie-1.fl0.io/",
+  baseURL: "http://localhost:3001",
 });
 
 //otra forma mas corta seria:
@@ -29,8 +29,8 @@ HopPassionClient.interceptors.response.use(
   },
   (error) => {
     if (error.response && error.response.status === 401) {
-      navigate("/")
-      store.dispatch(logout())
+      navigate("/");
+      store.dispatch(logout());
     }
     return Promise.reject(error);
   }
